Update to SpiderMonkey 87

Yes, you readied correctly. Mozilla creates mozjs tar one week prior release.

+++ mozjs/js/src/jsapi-tests/testArrayBufferView.cpp
@@ -76,14 +76,14 @@ static JSObject* CreateDataView(JSContext* cx) {
   return JS_NewDataView(cx, buffer, 0, 8);
 }
 
-template <JSObject* CreateTypedArray(JSContext* cx, uint32_t length),
+template <JSObject* CreateTypedArray(JSContext* cx, size_t length),
           size_t Length>
 static JSObject* Create(JSContext* cx) {
   return CreateTypedArray(cx, Length);
 }
 
 template <typename T, JSObject* CreateViewType(JSContext* cx),
-          JSObject* GetObjectAs(JSObject* obj, uint32_t* length,
+          JSObject* GetObjectAs(JSObject* obj, size_t* length,
                                 bool* isSharedMemory, T** data),
           js::Scalar::Type ExpectedType, uint32_t ExpectedLength,
           uint32_t ExpectedByteLength>
@@ -104,7 +104,7 @@ bool TestViewType(JSContext* cx) {
 
     T* data2;
     bool shared2;
-    uint32_t len;
+    size_t len;
     CHECK(obj == GetObjectAs(obj, &len, &shared2, &data2));
     CHECK(data1 == data2);
     CHECK(shared1 == shared2);
