+++ mozjs/js/src/jsapi-tests/testTypedArrays.cpp
@@ -111,12 +111,12 @@ BEGIN_TEST(testTypedArrays) {
 // Shared memory can only be mapped by a TypedArray by creating the
 // TypedArray with a SharedArrayBuffer explicitly, so no tests here.
 
-template <JSObject* Create(JSContext*, uint32_t), typename Element,
+template <JSObject* Create(JSContext*, size_t), typename Element,
           Element* GetData(JSObject*, bool* isShared,
                            const JS::AutoRequireNoGC&)>
 bool TestPlainTypedArray(JSContext* cx) {
   {
-    RootedObject notArray(cx, Create(cx, UINT32_MAX));
+    RootedObject notArray(cx, Create(cx, SIZE_MAX));
     CHECK(!notArray);
   }
